refactor: extract SettingsPanel appearance and advanced tabs into separate files
